Differential Alternative Splicing Event analysis

Use Limma, DESeq2, DoubleExpSeq and satuRn wrapper functions to test for differential Alternative Splice Events (ASEs)

ASE_limma(
  se,
  test_factor,
  test_nom,
  test_denom,
  batch1 = "",
  batch2 = "",
  IRmode = c("all", "annotated", "annotated_binary"),
  filter_antiover = TRUE,
  filter_antinear = FALSE
)

ASE_limma_timeseries(
  se,
  test_factor,
  batch1 = "",
  batch2 = "",
  degrees_of_freedom = 1,
  IRmode = c("all", "annotated", "annotated_binary"),
  filter_antiover = TRUE,
  filter_antinear = FALSE
)

ASE_DESeq(
  se,
  test_factor,
  test_nom,
  test_denom,
  batch1 = "",
  batch2 = "",
  n_threads = 1,
  IRmode = c("all", "annotated", "annotated_binary"),
  filter_antiover = TRUE,
  filter_antinear = FALSE
)

ASE_DoubleExpSeq(
  se,
  test_factor,
  test_nom,
  test_denom,
  IRmode = c("all", "annotated", "annotated_binary"),
  filter_antiover = TRUE,
  filter_antinear = FALSE
)

ASE_satuRn(
  se,
  test_factor,
  test_nom,
  test_denom,
  batch1 = "",
  batch2 = "",
  n_threads = 1,
  IRmode = c("all", "annotated", "annotated_binary"),
  filter_antiover = TRUE,
  filter_antinear = FALSE,
  filterByMinCPM = 0
)

Arguments

se

The NxtSE object created by makeSE(). To reduce runtime and avoid excessive multiple testing, consider filtering the object using applyFilters

test_factor

The column name in the sample annotation colData(se) that contains the desired variables to be contrasted. For ASE_limma_timeseries() and ASE_DESeq() (when test_nom and test_denom parameters are left blank), test_factor must contain numerical values representing the time variable.

test_nom

The nominator condition to test for differential ASE. Usually the "treatment" condition

test_denom

The denominator condition to test against for differential ASE. Usually the "control" condition

batch1, batch2

(Optional, limma and DESeq2 only) One or two condition types containing batch information to account for.

IRmode

(default all) Choose the approach to quantify IR events. Default all considers all introns as potentially retained, and calculates IR-ratio based on total splicing across the intron using the "SpliceOver" or "SpliceMax" approach (see collateData). Other options include annotated which calculates IR-ratios for annotated introns only, and annotated_binary which calculates PSI considering the "included" isoform as the IR-transcript, and the "excluded" transcript is quantified from splice counts only across the exact intron (but not that of overlapping introns). IR-ratio are denoted as "IR" events, whereas PSIs calculated using IR and intron-spliced binary alternatives are denoted as "RI" events.

filter_antiover, filter_antinear

Whether to remove novel IR events that overlap over or near anti-sense genes. Default will exclude antiover but not antinear introns. These are ignored if strand-specific RNA-seq protocols are used.

degrees_of_freedom

(default 1) The complexity of time series trends modeled by ASE_limma_timeseries. E.g., 1 will only model linear trends, whereas 2 extends the capacity for quadratic trends, 3 for cubic trends, etc.

n_threads

(DESeq2 only) How many threads to use for DESeq2 based analysis.

filterByMinCPM

(default 0) In ASE_satuRn(), Included/Excluded counts will be filtered using this value as the threshold prior to satuRn analysis. Filtering is performed using edgeR::filterByExpr() parsing this parameter into its min.count parameter.

Value

For all methods, a data.table containing the following:

EventName: The name of the ASE event. This identifies each ASE in downstream functions including makeMeanPSI, makeMatrix, and plotCoverage

EventType: The type of event. See details section above.

EventRegion: The genomic coordinates the event occupies. This spans the most upstream and most downstream splice junction involved in the ASE, and is use to guide the plotCoverage function.

flags: Indicates which isoforms are NMD substrates and/or which are formed by novel splicing only.

AvgPSI_nom, Avg_PSI_denom: the average percent spliced in / percent IR levels for the two conditions being contrasted. nom and denom in column names are replaced with the condition names. Note this is a geometric mean, based on the arithmetic mean of logit PSI values.

deltaPSI: The difference in PSI between the mean values of the two conditions.

limma specific output

logFC, AveExpr, t, P.Value, adj.P.Val, B: limma topTable columns of differential ASE. See limma::topTable for details.

inc/exc_(logFC, AveExpr, t, P.Value, adj.P.Val, B): limma results for differential testing for raw included / excluded counts only

DESeq2 specific output

baseMean, log2FoldChange, lfcSE, stat, pvalue, padj: DESeq2 results columns for differential ASE; see DESeq2::results for details.

inc/exc_(baseMean, log2FoldChange, lfcSE, stat, pvalue, padj): DESeq2 results for differential testing for raw included / excluded counts only

satuRn specific output

estimates, se, df, t, pval, regular_FDR: estimated log-odds ratio, standard error, degrees of freedom, (Wald) t statistic, nominal p-value and associated false discovery rate

empirical_pval, empirical_FDR: nominal p value and associated FDR computed by estimating the null distribution of the test statistic empirically (by satuRn).

DoubleExp specific output

MLE_nom, MLE_denom: Maximum likelihood expectation of PSI values for the denom in column names are replaced with the condition names

MLE_LFC: Log2-fold change of the MLE

P.Value, adj.P.Val: Nominal and BH-adjusted P values

n_eff: Number of effective samples (i.e. non-zero or non-unity PSI)

mDepth: Mean Depth of splice coverage in each of the two groups.

Dispersion_Reduced, Dispersion_Full: Dispersion values for reduced and full models. See DoubleExpSeq::DBGLM1 for details.

Details

Using limma, SpliceWiz models included and excluded counts as log-normal distributed, whereas using DESeq2, SpliceWiz models included and excluded counts as negative binomial distributed with dispersion shrinkage according to their mean count expressions. For limma and DESeq2, differential ASE are considered as the "interaction" between included and excluded splice counts for each sample. See this vignette for an explanation of how this is done.

SpliceWiz's limma wrapper implements an additional filter where ASEs with an average cpm values of either Included or Excluded counts are less than 1. DESeq2 has its own method for handling outliers, which seems to work well for handling situations where PSI ~= 0 or PSI ~= 1.

Time series are supported by SpliceWiz to a limited extent. Time series analysis can be performed via limma or DESeq2. For limma time-series analysis, use ASE_limma_timeseries(), specifying the test_factor as the column of numeric values containing time series data. For DESeq, time series differential analysis can be activated using the ASE_DESeq() function, again specifying test_factor as the column containing time series data (and leaving test_nom and test_denom parameters blank). See examples below.

Using DoubleExpSeq, included and excluded counts are modeled using the generalized beta prime distribution, using empirical Bayes shrinkage to estimate dispersion.

Using satuRn, included and excluded counts are modeled using the quasi-binomial distribution in a generalised linear model.

EventType are as follow:

IR = intron retention (IR-ratio) - all introns are considered

MXE = mutually exclusive exons

SE = skipped exons

AFE = alternate first exon

ALE = alternate last exon

A5SS = alternate 5'-splice site

A3SS = alternate 3'-splice site

RI = (known / annotated) intron retention (PSI).

NB: SpliceWiz measures intron retention events using two different approaches, the choice of which is left to the user - see ASE-methods:

IR (intron retention) events: considers all introns to be potentially retained. Given in most scenarios there may be uncertainty as to which of the many mutually-overlapping introns are spliced to produce the major isoform, SpliceWiz adopts the IRFinder approach by using the IR-ratio. The "included" isoform is the relative abundance of the IR-transcript, as approximated by the trimmed-mean depth of coverage across the intron (excluding outliers including exons of other transcripts, intronic elements such as snoRNAs, etc). The "excluded isoform" includes all spliced transcripts that contain an overlapping intron, as estimated via SpliceWiz's SpliceOver and IRFinder's SpliceMax methods - see collateData.

RI (annotated retained introns) considers only annotated retained introns, i.e., those annotated within the given reference. These are quantified using PSI, considering the included (IR-transcript) and excluded (splicing of the exact intron) as binary alternatives.

SpliceWiz considers "included" counts as those that represent abundance of the "included" isoform, whereas "excluded" counts represent the abundance of the "excluded" isoform. To allow comparison between modalities, SpliceWiz applies a convention whereby the "included" transcript is one where its splice junctions are by definition shorter than those of "excluded" transcripts. Specifically, this means the included / excluded isoforms are as follows:

EventType	Included	Excluded

IR or RI	Intron Retention	Spliced Intron

MXE	Upstream exon inclusion	Downstream exon inclusion

SE	Exon inclusion	Exon skipping

AFE	Downstream exon usage	Upstream exon usage

ALE	Upstream exon usage	Downstream exon usage

A5SS	Downstream 5'-SS	Upstream 5'-SS

A3SS	Upstream 3'-SS	Downstream 3'-SS

Functions

ASE_limma: Use limma to perform differential ASE analysis of a filtered NxtSE object

ASE_limma_timeseries: Use limma to perform differential ASE analysis of a filtered NxtSE object (time series)

ASE_DESeq: Use DESeq2 to perform differential ASE analysis of a filtered NxtSE object

ASE_DoubleExpSeq: Use DoubleExpSeq to perform differential ASE analysis of a filtered NxtSE object (uses double exponential beta-binomial model) to estimate group dispersions, followed by LRT

ASE_satuRn: Use satuRn to perform differential ASE analysis of a filtered NxtSE object

References

Ritchie ME, Phipson B, Wu D, Hu Y, Law CW, Shi W, Smyth GK (2015). 'limma powers differential expression analyses for RNA-sequencing and microarray studies.' Nucleic Acids Research, 43(7), e47. https://doi.org/10.1093/nar/gkv007

Love MI, Huber W, Anders S (2014). 'Moderated estimation of fold change and dispersion for RNA-seq data with DESeq2.' Genome Biology, 15, 550. https://doi.org/10.1186/s13059-014-0550-8

Ruddy S, Johnson M, Purdom E (2016). 'Shrinkage of dispersion parameters in the binomial family, with application to differential exon skipping.' Ann. Appl. Stat. 10(2): 690-725. https://doi.org/10.1214/15-AOAS871

Gilis J, Vitting-Seerup K, Van den Berge K, Clement L (2021). 'Scalable analysis of differential transcript usage for bulk and single-cell RNA-sequencing applications.' F1000Research 2021, 10:374. https://doi.org/10.12688/f1000research.51749.1

Examples

# Load the NxtSE object and set up the annotations
# - see ?makeSE on example code of generating this NxtSE object
se <- SpliceWiz_example_NxtSE()

colData(se)$treatment <- rep(c("A", "B"), each = 3)
colData(se)$replicate <- rep(c("P","Q","R"), 2)

# Limma analysis (counts modeled using log-normal distribution)

require("limma")
#> Loading required package: limma
res_limma <- ASE_limma(se, "treatment", "A", "B")
#> Nov 15 16:15:33 Performing limma contrast for included / excluded counts separately
#> Nov 15 16:15:34 Performing limma contrast for included / excluded counts together

# DoubleExpSeq analysis (counts modeled using beta binomial distribution)

require("DoubleExpSeq")
#> Loading required package: DoubleExpSeq
res_DES <- ASE_DoubleExpSeq(se, "treatment", "A", "B")
#> Nov 15 16:15:36 Running DoubleExpSeq::DBGLM1() on given data

# satuRn analysis (quasi binomial), 
#   Filtering counts using 1 count per million as threshold prior to analysis

require("satuRn")
#> Loading required package: satuRn
require("edgeR")
#> Loading required package: edgeR
#> Warning: there is no package called ‘edgeR’

res_sat <- ASE_satuRn(se, "treatment", "A", "B", filterByMinCPM = 1)
#> Nov 15 16:15:38 Using MulticoreParam 1 threads
#> Error: Nov 15 16:15:38 edgeR version 3.28.1 is not installed; and is required for this function

# DESeq2 analysis (counts modeled using negative binomial distribution)

require("DESeq2")
#> Loading required package: DESeq2
#> Loading required package: S4Vectors
#> Loading required package: stats4
#> Loading required package: BiocGenerics
#> 
#> Attaching package: ‘BiocGenerics’
#> The following object is masked from ‘package:limma’:
#> 
#>     plotMA
#> The following objects are masked from ‘package:stats’:
#> 
#>     IQR, mad, sd, var, xtabs
#> The following objects are masked from ‘package:base’:
#> 
#>     Filter, Find, Map, Position, Reduce, anyDuplicated, aperm, append,
#>     as.data.frame, basename, cbind, colnames, dirname, do.call,
#>     duplicated, eval, evalq, get, grep, grepl, intersect, is.unsorted,
#>     lapply, mapply, match, mget, order, paste, pmax, pmax.int, pmin,
#>     pmin.int, rank, rbind, rownames, sapply, setdiff, sort, table,
#>     tapply, union, unique, unsplit, which.max, which.min
#> 
#> Attaching package: ‘S4Vectors’
#> The following objects are masked from ‘package:base’:
#> 
#>     I, expand.grid, unname
#> Loading required package: IRanges
#> Loading required package: GenomicRanges
#> Loading required package: GenomeInfoDb
#> Loading required package: SummarizedExperiment
#> Loading required package: MatrixGenerics
#> Loading required package: matrixStats
#> 
#> Attaching package: ‘MatrixGenerics’
#> The following objects are masked from ‘package:matrixStats’:
#> 
#>     colAlls, colAnyNAs, colAnys, colAvgsPerRowSet, colCollapse,
#>     colCounts, colCummaxs, colCummins, colCumprods, colCumsums,
#>     colDiffs, colIQRDiffs, colIQRs, colLogSumExps, colMadDiffs,
#>     colMads, colMaxs, colMeans2, colMedians, colMins, colOrderStats,
#>     colProds, colQuantiles, colRanges, colRanks, colSdDiffs, colSds,
#>     colSums2, colTabulates, colVarDiffs, colVars, colWeightedMads,
#>     colWeightedMeans, colWeightedMedians, colWeightedSds,
#>     colWeightedVars, rowAlls, rowAnyNAs, rowAnys, rowAvgsPerColSet,
#>     rowCollapse, rowCounts, rowCummaxs, rowCummins, rowCumprods,
#>     rowCumsums, rowDiffs, rowIQRDiffs, rowIQRs, rowLogSumExps,
#>     rowMadDiffs, rowMads, rowMaxs, rowMeans2, rowMedians, rowMins,
#>     rowOrderStats, rowProds, rowQuantiles, rowRanges, rowRanks,
#>     rowSdDiffs, rowSds, rowSums2, rowTabulates, rowVarDiffs, rowVars,
#>     rowWeightedMads, rowWeightedMeans, rowWeightedMedians,
#>     rowWeightedSds, rowWeightedVars
#> Loading required package: Biobase
#> Welcome to Bioconductor
#> 
#>     Vignettes contain introductory material; view with
#>     'browseVignettes()'. To cite Bioconductor, see
#>     'citation("Biobase")', and for packages 'citation("pkgname")'.
#> 
#> Attaching package: ‘Biobase’
#> The following object is masked from ‘package:MatrixGenerics’:
#> 
#>     rowMedians
#> The following objects are masked from ‘package:matrixStats’:
#> 
#>     anyMissing, rowMedians
res_DESeq <- ASE_DESeq(se, "treatment", "A", "B")
#> Nov 15 16:15:39 Using MulticoreParam 1 threads
#> Nov 15 16:15:39 Performing DESeq2 contrast for included / excluded counts separately
#> converting counts to integer mode
#> ASE_DESeq: Profiling expression of Included and Excluded counts
#> estimating size factors
#> estimating dispersions
#> gene-wise dispersion estimates: 1 workers
#> mean-dispersion relationship
#> final dispersion estimates, fitting model and testing: 1 workers
#> Nov 15 16:15:42 Performing DESeq2 contrast for included / excluded counts separately
#> ASE_DESeq: Profiling differential ASE
#> using pre-existing size factors
#> estimating dispersions
#> gene-wise dispersion estimates: 1 workers
#> mean-dispersion relationship
#> final dispersion estimates, fitting model and testing: 1 workers

# Time series examples

colData(se)$timepoint <- rep(c(1,2,3), each = 2)
colData(se)$batch <- rep(c("1", "2"), 3)

res_limma_timeseries <- ASE_limma_timeseries(se, "timepoint")
#> Nov 15 16:15:45 Performing limma contrast for included / excluded counts separately
#> Nov 15 16:15:46 Performing limma contrast for included / excluded counts together
res_DESeq_timeseries <- ASE_DESeq(se, "timepoint")
#> Nov 15 16:15:49 Using MulticoreParam 1 threads
#> Nov 15 16:15:49 Performing DESeq2 contrast for included / excluded counts separately
#> converting counts to integer mode
#>   the design formula contains one or more numeric variables with integer values,
#>   specifying a model with increasing fold change for higher values.
#>   did you mean for this to be a factor? if so, first convert
#>   this variable to a factor using the factor() function
#> ASE_DESeq: Profiling expression of Included and Excluded counts
#> estimating size factors
#> estimating dispersions
#> gene-wise dispersion estimates: 1 workers
#> mean-dispersion relationship
#> final dispersion estimates, fitting model and testing: 1 workers
#> Nov 15 16:15:51 Performing DESeq2 contrast for included / excluded counts separately
#>   the design formula contains one or more numeric variables with integer values,
#>   specifying a model with increasing fold change for higher values.
#>   did you mean for this to be a factor? if so, first convert
#>   this variable to a factor using the factor() function
#> ASE_DESeq: Profiling differential ASE
#> using pre-existing size factors
#> estimating dispersions
#> gene-wise dispersion estimates: 1 workers
#> mean-dispersion relationship
#> final dispersion estimates, fitting model and testing: 1 workers