The NxtSE class

The NxtSE class inherits from the SummarizedExperiment class and is constructed using makeSE. NxtSE extends SummarizedExperiment by housing additional assays pertaining to IR and splice junction counts.

NxtSE(...)

# S4 method for NxtSE
up_inc(x, withDimnames = TRUE, ...)

# S4 method for NxtSE
down_inc(x, withDimnames = TRUE, ...)

# S4 method for NxtSE
up_exc(x, withDimnames = TRUE, ...)

# S4 method for NxtSE
down_exc(x, withDimnames = TRUE, ...)

# S4 method for NxtSE
covfile(x, withDimnames = TRUE, ...)

# S4 method for NxtSE
sampleQC(x, withDimnames = TRUE, ...)

# S4 method for NxtSE
ref(x, withDimnames = TRUE, ...)

# S4 method for NxtSE
junc_PSI(x, withDimnames = TRUE, ...)

# S4 method for NxtSE
junc_counts(x, withDimnames = TRUE, ...)

# S4 method for NxtSE
junc_gr(x, withDimnames = TRUE, ...)

# S4 method for NxtSE
realize_NxtSE(x, includeJunctions = FALSE, withDimnames = TRUE, ...)

# S4 method for NxtSE
up_inc(x, withDimnames = TRUE) <- value

# S4 method for NxtSE
down_inc(x, withDimnames = TRUE) <- value

# S4 method for NxtSE
up_exc(x, withDimnames = TRUE) <- value

# S4 method for NxtSE
down_exc(x, withDimnames = TRUE) <- value

# S4 method for NxtSE
covfile(x, withDimnames = TRUE) <- value

# S4 method for NxtSE
sampleQC(x, withDimnames = TRUE) <- value

# S4 method for NxtSE,ANY,ANY,ANY
[(x, i, j, ..., drop = TRUE)

# S4 method for NxtSE,ANY,ANY,NxtSE
[(x, i, j, ...) <- value

# S4 method for NxtSE
cbind(..., deparse.level = 1)

# S4 method for NxtSE
rbind(..., deparse.level = 1)

Arguments

...

In NxtSE(), additional arguments to be passed onto SummarizedExperiment()

x

A NxtSE object

withDimnames

(default TRUE) Whether exported assays should be supplied with row and column names of the NxtSE object. See SummarizedExperiment

includeJunctions

When realizing a NxtSE object, include whether junction counts and PSIs should be realized into memory. Not recommended for general use, as they are only used for coverage plots.

value

The value to replace. Must be a matrix for the up_inc<-, down_inc<-, up_exc<- and down_exc<- replacers, and a character vector for covfile<-

i, j

Row and column subscripts to subset a NxtSE object.

drop

A logical(1), ignored by these methods.

deparse.level

See base::cbind for a description of this argument.

Value

See Functions section (below) for details

Functions

NxtSE: Constructor function for NxtSE; akin to SummarizedExperiment(...)

up_inc,NxtSE-method: Gets upstream included events (SE/MXE), or upstream exon-intron spanning reads (IR)

down_inc,NxtSE-method: Gets downstream included events (SE/MXE), or downstream exon-intron spanning reads (IR)

up_exc,NxtSE-method: Gets upstream excluded events (MXE only)

down_exc,NxtSE-method: Gets downstream excluded events (MXE only)

covfile,NxtSE-method: Gets a named vector with the paths to the corresponding COV files

sampleQC,NxtSE-method: Gets a data frame with the QC parameters of the samples

ref,NxtSE-method: Retrieves a list of annotation data associated with this NxtSE object; primarily used in plotCoverage()

junc_PSI,NxtSE-method: Getter for junction PSI DelayedMatrix; primarily used in plotCoverage()

junc_counts,NxtSE-method: Getter for junction counts DelayedMatrix; primarily used in plotCoverage()

junc_gr,NxtSE-method: Getter for junction GenomicRanges coordinates; primarily used in plotCoverage()

realize_NxtSE,NxtSE-method: Converts all DelayedMatrix assays as matrices (i.e. performs all delayed calculation and loads resulting object to RAM)

up_inc<-,NxtSE-method: Sets upstream included events (SE/MXE), or upstream exon-intron spanning reads (IR)

down_inc<-,NxtSE-method: Sets downstream included events (SE/MXE), or downstream exon-intron spanning reads (IR)

up_exc<-,NxtSE-method: Sets upstream excluded events (MXE only)

down_exc<-,NxtSE-method: Sets downstream excluded events (MXE only)

covfile<-,NxtSE-method: Sets the paths to the corresponding COV files

sampleQC<-,NxtSE-method: Sets the values in the data frame containing sample QC

[,NxtSE,ANY,ANY,ANY-method: Subsets a NxtSE object

[<-,NxtSE,ANY,ANY,NxtSE-method: Sets a subsetted NxtSE object

cbind,NxtSE-method: Combines two NxtSE objects (by samples - columns)

rbind,NxtSE-method: Combines two NxtSE objects (by AS/IR events - rows)

Examples

# Run the full pipeline to generate a NxtSE object:

buildRef(
    reference_path = file.path(tempdir(), "Reference"),
    fasta = chrZ_genome(), 
    gtf = chrZ_gtf()
)
#> Nov 15 16:16:17 SpliceWiz reference already exists in given directory
#> NULL

bams <- SpliceWiz_example_bams()
processBAM(bams$path, bams$sample,
  reference_path = file.path(tempdir(), "Reference"),
  output_path = file.path(tempdir(), "SpliceWiz_Output")
)
#> Nov 15 16:16:20 Running SpliceWiz processBAM
#> Nov 15 16:16:20 Using MulticoreParam 1 threads
#> Nov 15 16:16:20 SpliceWiz processBAM processed Rtmpd5xBOp/02H003.bam
#> Nov 15 16:16:20 SpliceWiz processBAM processed Rtmpd5xBOp/02H025.bam
#> Nov 15 16:16:21 SpliceWiz processBAM processed Rtmpd5xBOp/02H026.bam
#> Nov 15 16:16:21 SpliceWiz processBAM processed Rtmpd5xBOp/02H033.bam
#> Nov 15 16:16:21 SpliceWiz processBAM processed Rtmpd5xBOp/02H043.bam
#> Nov 15 16:16:22 SpliceWiz processBAM processed Rtmpd5xBOp/02H046.bam

expr <- findSpliceWizOutput(file.path(tempdir(), "SpliceWiz_Output"))
collateData(expr, 
  reference_path = file.path(tempdir(), "Reference"),
  output_path = file.path(tempdir(), "Collated_output")
)
#> Nov 15 16:16:22 Using MulticoreParam 1 threads
#> Nov 15 16:16:22 Validating Experiment; checking COV files...
#> Nov 15 16:16:22 Compiling Sample Stats
#> Nov 15 16:16:22 Compiling Junction List...
#> merging...
#> done
#> Nov 15 16:16:23 Compiling Junction Stats...
#> merging...
#> done
#> Nov 15 16:16:27 Compiling Intron Retention List...
#> done
#> Nov 15 16:16:28 Tidying up splice junctions and intron retentions...
#> ...annotating splice junctions
#> ...grouping splice junctions
#> ...grouping introns
#> ...loading splice events
#> ...compiling rowEvents
#> done
#> Nov 15 16:16:33 Generating NxtSE assays
#> Nov 15 16:16:33 Using MulticoreParam 1 threads
#>   |                                                                              |                                                                      |   0%  |                                                                              |============                                                          |  17%  |                                                                              |=======================                                               |  33%  |                                                                              |===================================                                   |  50%  |                                                                              |===============================================                       |  67%  |                                                                              |==========================================================            |  83%  |                                                                              |======================================================================| 100%
#> 
#> Nov 15 16:16:42 Building Final NxtSE Object
#> Nov 15 16:16:44 SpliceWiz (NxtSE) Collation Finished

se <- makeSE(collate_path = file.path(tempdir(), "Collated_output"))
#> Nov 15 16:16:44 Loading NxtSE object from file...
#> done
#> Nov 15 16:16:44 Removing overlapping introns...
#> Nov 15 16:16:44 Iterating through IR events to determine introns of main isoforms
#> Nov 15 16:16:45 Iteration 1
#> Nov 15 16:16:45 Iteration 2

# Coerce NxtSE -> SummarizedExperiment
se_raw <- as(se, "SummarizedExperiment")

# Coerce SummarizedExperiment -> NxtSE
se_NxtSE <- as(se_raw, "NxtSE")
identical(se, se_NxtSE) # Returns TRUE
#> [1] TRUE

# Get Junction reads of SE / MXE and spans-reads of IR events
up_inc(se)
#> <90 x 6> matrix of class DelayedMatrix and type "double":
#>                                      02H003 02H025 02H026 02H033 02H043 02H046
#>  TRA2B/ENST00000453386_Intron8/clean     14      4     16     36      7     27
#>  TRA2B/ENST00000453386_Intron7/clean     15      1     53     21      7     29
#>  TRA2B/ENST00000453386_Intron6/clean     18      5     71     22      5     80
#>  TRA2B/ENST00000453386_Intron5/clean      1      2     20      4      1     24
#>  TRA2B/ENST00000453386_Intron4/clean     51      6     56     48     14     75
#>                                  ...      .      .      .      .      .      .
#> RI:SRSF1-204-exon3;SRSF1-201-intron3    141    101     37    103    221     73
#> RI:SRSF2-202-exon3;SRSF2-209-intron1   1947   1291   1697   1569   1134   1102
#> RI:SRSF2-203-exon2;SRSF2-206-intron2    529    219    187    251    330    342
#> RI:SRSF2-203-exon2;SRSF2-202-intron2    499    228    159    213    238    305
#> RI:SRSF2-203-exon2;SRSF2-206-intron1    499    228    159    213    238    305
down_inc(se)
#> <90 x 6> matrix of class DelayedMatrix and type "double":
#>                                      02H003 02H025 02H026 02H033 02H043 02H046
#>  TRA2B/ENST00000453386_Intron8/clean     21     13     63     72     19     52
#>  TRA2B/ENST00000453386_Intron7/clean     54     29    141     80     30    100
#>  TRA2B/ENST00000453386_Intron6/clean     37     27     91     53     14    102
#>  TRA2B/ENST00000453386_Intron5/clean      3      6     42     28      3     29
#>  TRA2B/ENST00000453386_Intron4/clean     67     41     97     72     33    114
#>                                  ...      .      .      .      .      .      .
#> RI:SRSF1-204-exon3;SRSF1-201-intron3    174    106     66    137    260    109
#> RI:SRSF2-202-exon3;SRSF2-209-intron1   2207   1456   2142   1995   1585   1564
#> RI:SRSF2-203-exon2;SRSF2-206-intron2    845    475    356    468    593    514
#> RI:SRSF2-203-exon2;SRSF2-202-intron2    845    475    356    468    593    514
#> RI:SRSF2-203-exon2;SRSF2-206-intron1    571    213    176    242    359    377
up_exc(se)
#> <3 x 6> matrix of class DelayedMatrix and type "double":
#>                                     02H003 02H025 02H026 02H033 02H043 02H046
#> MXE:TRA2A-212-exon2;TRA2A-203-exon2    126     75     45     95     35     94
#> MXE:TRA2A-212-exon2;TRA2A-208-exon2    126     75     45     95     35     94
#> MXE:TRA2A-212-exon2;TRA2A-211-exon2    126     75     45     95     35     94
down_exc(se)
#> <3 x 6> matrix of class DelayedMatrix and type "double":
#>                                     02H003 02H025 02H026 02H033 02H043 02H046
#> MXE:TRA2A-212-exon2;TRA2A-203-exon2      8      4      6     10      0      3
#> MXE:TRA2A-212-exon2;TRA2A-208-exon2     70     35     33     58     22     69
#> MXE:TRA2A-212-exon2;TRA2A-211-exon2     32     23     13     12     11     17

# Get list of available coverage files
covfile(se)
#>                                                                                            02H003 
#> "/private/var/folders/xy/6ybn2drx76lblc_fqhfwt8qw0000gr/T/Rtmpd5xBOp/SpliceWiz_Output/02H003.cov" 
#>                                                                                            02H025 
#> "/private/var/folders/xy/6ybn2drx76lblc_fqhfwt8qw0000gr/T/Rtmpd5xBOp/SpliceWiz_Output/02H025.cov" 
#>                                                                                            02H026 
#> "/private/var/folders/xy/6ybn2drx76lblc_fqhfwt8qw0000gr/T/Rtmpd5xBOp/SpliceWiz_Output/02H026.cov" 
#>                                                                                            02H033 
#> "/private/var/folders/xy/6ybn2drx76lblc_fqhfwt8qw0000gr/T/Rtmpd5xBOp/SpliceWiz_Output/02H033.cov" 
#>                                                                                            02H043 
#> "/private/var/folders/xy/6ybn2drx76lblc_fqhfwt8qw0000gr/T/Rtmpd5xBOp/SpliceWiz_Output/02H043.cov" 
#>                                                                                            02H046 
#> "/private/var/folders/xy/6ybn2drx76lblc_fqhfwt8qw0000gr/T/Rtmpd5xBOp/SpliceWiz_Output/02H046.cov" 

# Get sample QC information
sampleQC(se)
#>                                     path paired strand  depth mean_frag_size
#> 02H003 ../SpliceWiz_Output/02H003.txt.gz   TRUE     -1 106256       149.3108
#> 02H025 ../SpliceWiz_Output/02H025.txt.gz   TRUE     -1  84872       151.5914
#> 02H026 ../SpliceWiz_Output/02H026.txt.gz   TRUE     -1 131514       150.3238
#> 02H033 ../SpliceWiz_Output/02H033.txt.gz   TRUE     -1 129013       153.9754
#> 02H043 ../SpliceWiz_Output/02H043.txt.gz   TRUE     -1  80254       155.4004
#> 02H046 ../SpliceWiz_Output/02H046.txt.gz   TRUE     -1 120174       151.2840
#>        directionality_strength Intergenic_Fraction rRNA_Fraction
#> 02H003                    9850                   0             0
#> 02H025                    9848                   0             0
#> 02H026                    9848                   0             0
#> 02H033                    9855                   0             0
#> 02H043                    9841                   0             0
#> 02H046                    9857                   0             0
#>        NonPolyA_Fraction Mitochondrial_Fraction Unanno_Jn_Fraction
#> 02H003                 0                      0        0.013449472
#> 02H025                 0                      0        0.011203515
#> 02H026                 0                      0        0.007282382
#> 02H033                 0                      0        0.006310241
#> 02H043                 0                      0        0.005871040
#> 02H046                 0                      0        0.009819533
#>        NMD_Jn_Fraction Fraction_Splice_Reads Fraction_Span_Reads IRBurden_clean
#> 02H003     0.019509970             0.3506908           0.1735243     0.05796583
#> 02H025     0.019360162             0.3712414           0.1591691     0.03963531
#> 02H026     0.004298323             0.3980261           0.1615037     0.02576549
#> 02H033     0.007705562             0.4003550           0.1734089     0.03090263
#> 02H043     0.028069104             0.3671219           0.1807511     0.03762138
#> 02H046     0.010542303             0.3725598           0.1603508     0.05099352
#>        IRBurden_exitrons IRBurden_clean_unstranded IRBurden_exitrons_unstranded
#> 02H003         0.1291810                0.03844293                    0.1296020
#> 02H025         0.1192351                0.02730399                    0.1194005
#> 02H026         0.1231055                0.02010410                    0.1231939
#> 02H033         0.1378958                0.02300500                    0.1380215
#> 02H043         0.1392935                0.02286515                    0.1397365
#> 02H046         0.1132136                0.03359077                    0.1133852
#>        IRBurden_antisense
#> 02H003        0.021393335
#> 02H025        0.013191532
#> 02H026        0.006007119
#> 02H033        0.008453742
#> 02H043        0.015837096
#> 02H046        0.019043703

# Get resource data (used internally for plotCoverage())
cov_data <- ref(se)
names(cov_data)
#> [1] "seqInfo"        "gene_list"      "elem.DT"        "transcripts.DT"

# Subset functions
se_by_samples <- se[,1:3]
se_by_events <- se[1:10,]
se_by_rowData <- subset(se, EventType == "IR")

# Cbind (bind event_identical NxtSE by samples)
se_by_samples_1 <- se[,1:3]
se_by_samples_2 <- se[,4:6]
se_cbind <- cbind(se_by_samples_1, se_by_samples_2)
identical(se, se_cbind) # should return TRUE
#> [1] FALSE

# Rbind (bind sample_identical NxtSE by events)
se_IR <- subset(se, EventType == "IR")
se_SE <- subset(se, EventType == "SE")
se_IRSE <- rbind(se_IR, se_SE)
identical(se_IRSE, subset(se, EventType %in% c("IR", "SE"))) # TRUE
#> [1] FALSE

# Convert HDF5-based NxtSE to in-memory se
# makeSE() creates a HDF5-based NxtSE object where all assay data is stored
# as an h5 file instead of in-memory. All operations are performed as
# delayed operations as per DelayedArray package.
# To realize the NxtSE object as an in-memory object, use:

se_real <- realize_NxtSE(se)
identical(se, se_real) # should return FALSE
#> [1] FALSE

# To check the difference, run:
class(up_inc(se))
#> [1] "DelayedMatrix"
#> attr(,"package")
#> [1] "DelayedArray"
class(up_inc(se_real))
#> [1] "matrix" "array"