Filtering for IR and Alternative Splicing Events

These function implements filtering of alternative splicing events, based on customisable criteria. See ASEFilter for details on how to construct SpliceWiz filters

getDefaultFilters()

applyFilters(se, filters = getDefaultFilters())

runFilter(se, filterObj)

Arguments

se

the NxtSE object to filter

filters

A vector or list of one or more ASEFilter objects. If left blank, the SpliceWiz default filters will be used.

filterObj

A single ASEFilter object.

Value

For runFilter and applyFilters: a vector of type logical, representing the rows of NxtSE that should be kept.

For getDefaultFilters: returns a list of default recommended filters that should be parsed into applyFilters.

Details

We highly recommend using the default filters, which are as follows:

(1) Depth filter of 20,

(2) Participation filter requiring 70% coverage in IR events.

(3) Participation filter requiring 40% coverage in SE, A5SS and A3SS events (i.e. Included + Excluded isoforms must cover at least 40% of all junction events across the given region)

(4) Consistency filter requring log difference of 2 (for skipped exon and mutually exclusive exon events, each junction must comprise at least 1/(2^2) = 1/4 of all reads associated with each isoform). For retained introns, the exon-intron overhangs must not differ by 1/4

(5) Terminus filter: In alternate first exons, the splice junction must not be shared with another transcript for which it is not its first intron. For alternative last exons, the splice junction must not be shared with another transcript for which it is not its last intron

(6) ExclusiveMXE filter: For MXE events, the two alternate casette exons must not overlap in their genomic regions

In all data-based filters, we require at least 80% samples (pcTRUE = 80) to pass this filters from the entire dataset (minCond = -1).

Events with event read depth (reads supporting either included or excluded isoforms) lower than 5 (minDepth = 5) are not assessed in filter #2, and in #3 and #4 this threshold is (minDepth = 20).

For an explanation of the various parameters mentioned here, see ASEFilter

Functions

getDefaultFilters: Returns a vector of recommended default SpliceWiz filters

applyFilters: Run a vector or list of ASEFilter objects on a NxtSE object

runFilter: Run a single filter on a NxtSE object

See also

Examples

# see ?makeSE on example code of how this object was generated

se <- SpliceWiz_example_NxtSE()

# Get the list of SpliceWiz recommended filters

filters <- getDefaultFilters()

# View a description of what these filters do:

filters
#> [[1]]
#> ASEFilter Class: Data	Type: Depth
#> EventTypes: IR MXE SE A3SS A5SS AFE ALE RI
#> Filter must be passed in at least 80.0 percent of all samples
#> Minimum Event Depth: 20
#> Event Depth refers to number of aligned splice reads plus effective depth of their introns
#> 
#> [[2]]
#> ASEFilter Class: Data	Type: Participation
#> EventTypes: IR RI
#> Filter must be passed in at least 80.0 percent of all samples
#> Minimum Coverage (Participation) Percentage: 70.0
#> In retained introns, participation refers to the proportion of the measured intron that is covered by at least 1 alignment
#> Event Depth below 5 are ignored
#> Event Depth refers to number of aligned splice reads plus effective depth of their introns
#> 
#> [[3]]
#> ASEFilter Class: Data	Type: Participation
#> EventTypes: SE A5SS A3SS
#> Filter must be passed in at least 80.0 percent of all samples
#> Minimum Coverage (Participation) Percentage: 40.0
#> In splice events, participation refers to the proportion of junction reads that belong to either the included or excluded isoforms of the given event
#> Event Depth below 20 are ignored
#> Event Depth refers to number of aligned splice reads plus effective depth of their introns
#> 
#> [[4]]
#> ASEFilter Class: Data	Type: Consistency
#> EventTypes: MXE SE RI
#> Filter must be passed in at least 80.0 percent of all samples
#> Sub-event minimum proportion: 0.2 of total isoform splice count
#> In retained introns, sub-events refer to 5'- and 3'-exon-intron spanning alignments
#> In mutually exclusive exons and skipped exons, sub-events refer to the upstream and downstream splice alignments
#> Event Depth below 20 are ignored
#> Event Depth refers to number of aligned splice reads plus effective depth of their introns
#> 
#> [[5]]
#> ASEFilter Class: Annotation	Type: Terminus
#> EventTypes: IR MXE SE A3SS A5SS AFE ALE RI
#> For alternative first / last exons, events overlapping with non-first/last introns are removed
#> [[6]]
#> ASEFilter Class: Annotation	Type: ExclusiveMXE
#> EventTypes: IR MXE SE A3SS A5SS AFE ALE RI
#> MXE events with overlapping casette exons are removed

# Filter the NxtSE using the first default filter ("Depth")

se.depthfilter <- se[runFilter(se, filters[[1]]), ]
#> Running Depth filter

# Filter the NxtSE using all four default filters

se.defaultFiltered <- se[applyFilters(se, getDefaultFilters()), ]
#> Running Depth filter
#> Running Participation filter
#> Running Participation filter
#> Running Consistency filter
#> Running Terminus filter
#> Running ExclusiveMXE filter