SpliceWiz Example BAMs and NxtSE Experiment Object

SpliceWiz_example_bams() is a wrapper function to obtain and make a local copy of 6 example files provided by the NxtIRFdata companion package to demonstrate the use of SpliceWiz. See NxtIRFdata::example_bams for a description of the provided BAM files.

SpliceWiz_example_NxtSE() retrieves a ready-made functioning NxtSE object. The steps to reproduce this object is shown in the example code in makeSE

SpliceWiz_example_bams()

SpliceWiz_example_NxtSE(novelSplicing = FALSE)

Arguments

novelSplicing

Whether to import an example NxtSE with novel splice event discovery.

Value

In SpliceWiz_example_bams(): returns a 2-column data frame containing sample names and BAM paths of the example dataset.

In SpliceWiz_example_NxtSE(): returns a NxtSE object.

Functions

SpliceWiz_example_bams: Returns a 2-column data frame, containing sample names and sample paths (in tempdir()) of example BAM files

SpliceWiz_example_NxtSE: Returns a (in-memory / realized) NxtSE object that was pre-generated using the SpliceWiz example reference and example BAM files

References

Generation of the mappability files was performed using SpliceWiz using a method analogous to that described in:

Middleton R, Gao D, Thomas A, Singh B, Au A, Wong JJ, Bomane A, Cosson B, Eyras E, Rasko JE, Ritchie W. IRFinder: assessing the impact of intron retention on mammalian gene expression. Genome Biol. 2017 Mar 15;18(1):51. doi:10.1186/s13059-017-1184-4

See also

Examples

# returns a data frame with the first column as sample names, and the
# second column as BAM paths

SpliceWiz_example_bams()
#>   sample
#> 1 02H003
#> 2 02H025
#> 3 02H026
#> 4 02H033
#> 5 02H043
#> 6 02H046
#>                                                                      path
#> 1 /var/folders/xy/6ybn2drx76lblc_fqhfwt8qw0000gr/T//Rtmpd5xBOp/02H003.bam
#> 2 /var/folders/xy/6ybn2drx76lblc_fqhfwt8qw0000gr/T//Rtmpd5xBOp/02H025.bam
#> 3 /var/folders/xy/6ybn2drx76lblc_fqhfwt8qw0000gr/T//Rtmpd5xBOp/02H026.bam
#> 4 /var/folders/xy/6ybn2drx76lblc_fqhfwt8qw0000gr/T//Rtmpd5xBOp/02H033.bam
#> 5 /var/folders/xy/6ybn2drx76lblc_fqhfwt8qw0000gr/T//Rtmpd5xBOp/02H043.bam
#> 6 /var/folders/xy/6ybn2drx76lblc_fqhfwt8qw0000gr/T//Rtmpd5xBOp/02H046.bam

# Returns a NxtSE object created by the example bams aligned to the
# mock NxtSE reference

se <- SpliceWiz_example_NxtSE()