Runs the OpenMP/C++ based SpliceWiz algorithm

These function calls the SpliceWiz C++ routine on one or more BAM files.

The routine is an improved version over the original IRFinder, with OpenMP-based multi-threading and the production of compact "COV" files to record alignment coverage. A SpliceWiz reference built using Build-Reference-methods is required.

After processBAM() is run, users should call collateData to collate individual outputs into an experiment / dataset.

BAM2COV creates COV files from BAM files without running processBAM().

See details for performance info.

BAM2COV(
  bamfiles = "./Unsorted.bam",
  sample_names = "sample1",
  output_path = "./cov_folder",
  n_threads = 1,
  useOpenMP = TRUE,
  overwrite = FALSE,
  verbose = FALSE,
  multiRead = FALSE
)

processBAM(
  bamfiles = "./Unsorted.bam",
  sample_names = "sample1",
  reference_path = "./Reference",
  output_path = "./SpliceWiz_Output",
  n_threads = 1,
  useOpenMP = TRUE,
  overwrite = FALSE,
  run_featureCounts = FALSE,
  verbose = FALSE,
  multiRead = FALSE
)

Arguments

bamfiles

A vector containing file paths of 1 or more BAM files

sample_names

The sample names of the given BAM files. Must be a vector of the same length as bamfiles

output_path

The output directory of this function

n_threads

(default 1) The number of threads to use. See details.

useOpenMP

(default TRUE) Whether to use OpenMP. If set to FALSE, BiocParallel will be used if n_threads is set

overwrite

(default FALSE) If output files already exist, will not attempt to re-run. If run_featureCounts is TRUE, will not overwrite gene counts of previous run unless overwrite is TRUE.

verbose

(default FALSE) Set to TRUE to allow processBAM() to output progress bars and messages

multiRead

(default FALSE) Whether to use multiple threads to read files. Set to TRUE to use multiple threads (improves performance if files are already cached into memory.

reference_path

The directory containing the SpliceWiz reference

run_featureCounts

(default FALSE) Whether this function will run Rsubread::featureCounts on the BAM files after counting spliced reads. If so, the output will be saved to "main.FC.Rds in the output_path directory as a list object.

Value

processBAM() output will be saved to output_path. Output files will be named using the given sample names.

sample.txt.gz: The main output file containing the quantitation of IR and splice junctions, as well as QC information

sample.cov: Contains coverage information in compressed binary. See getCoverage

main.FC.Rds: A single file containing gene counts for the whole dataset (only if run_featureCounts == TRUE)

Details

Typical run-times for a 100-million paired-end alignment BAM file takes 10 minutes using a single core. Using 8 threads, the runtime is approximately 2-5 minutes, depending on your system's file input / output speeds. Approximately 10 Gb of RAM is used when OpenMP is used. If OpenMP is not used (see below), this memory usage is multiplied across the number of processor threads (i.e. 40 Gb if n_threads = 4).

OpenMP is natively available to Linux / Windows compilers, and OpenMP will be used if useOpenMP is set to TRUE, using multiple threads to process each BAM file. On Macs, if OpenMP is not available at compilation, BiocParallel will be used, processing BAM files simultaneously, with one BAM file per thread.

Functions

BAM2COV: Converts BAM files to COV files without running processBAM()

processBAM: Processes BAM files. Requires a SpliceWiz reference generated by buildRef()

See also

Examples

# Run BAM2COV, which only produces COV files but does not run `processBAM()`:

bams <- SpliceWiz_example_bams()

BAM2COV(bams$path, bams$sample,
  output_path = file.path(tempdir(), "SpliceWiz_Output"),
  n_threads = 2, overwrite = TRUE
)
#> Nov 15 16:17:13 Running BAM2COV
#> Nov 15 16:17:13 Using MulticoreParam 2 threads

# Run processBAM(), which produces:
# - text output of intron coverage and spliced read counts
# - COV files which record read coverages

example_ref <- file.path(tempdir(), "Reference")

buildRef(
    reference_path = example_ref,
    fasta = chrZ_genome(),
    gtf = chrZ_gtf()
)
#> Nov 15 16:17:15 SpliceWiz reference already exists in given directory
#> NULL

bams <- SpliceWiz_example_bams()

processBAM(bams$path, bams$sample,
  reference_path = file.path(tempdir(), "Reference"),
  output_path = file.path(tempdir(), "SpliceWiz_Output"),
  n_threads = 2
)
#> Nov 15 16:17:15 processBAM has already been run on given BAM files